Create make_weekday_sequence.py
Created the tk_make_weekday_sequence. It is slightly different and more encompassing than the R version in that it can adjust to alternative weekday schedules. For example, most of the Middle East, including Israel, works Sunday to Thursday. This enables users from those countries to easily generate this alternative without having to shift things.