Update lags.py
I have revised this function to be more flexible. The former version only enables an integer or list that behaved like a range. This was fine, but a more Pythonic pattern would be: integer, tuple, list. Where the integer adds a single lag at the prescribed integer, the tuple is simply a range, so (1,7) adds lag1 to lag7, and importantly, what was missed and is also not available in the R version of timetk are specific lags at different values, which is important for reasons the group knows related to diagnosing autocorrelation, etc. So a list, say [1,7] includes lag1 and lag7 but no lags between. The list can take any integer values, so it could be [1, 4, 7, 14] and you just get four lags lag1, lag4, lag7, and lag14 returned.