Create fourier.py
This adds the fourier transform in a very similar way to the R implementation. I saw the slack messages that indicated a desire to not have too many dependencies, I agree so rather than used scipy's fft for fourier I used radians for sin and cos and avoided creating another dependency (for now).